+++ src/main/java/appFacturacion/Corporativo.java
@@ -8,11 +8,13 @@ public class Corporativo extends Cliente {
     private String telefono;
     private String email;
 
-    public Corporativo(String telefono, String email, String Nombre, String Sexo, Date fecha, DetalleOrden detalleOrden, Pago pago, Cliente client, String producto, double Precio, double cantidad, ArrayList<Double> muestra) {
+    public Corporativo(String telefono, String email, String Nombre, String Sexo, Date fecha, DetalleOrden detalleOrden, Pago pago, Cliente client, String producto, float Precio, float cantidad, ArrayList<Double> muestra) {
         super(Nombre, Sexo, fecha, detalleOrden, pago, client, producto, Precio, cantidad, muestra);
         this.telefono = telefono;
         this.email = email;
     }
+
+    
 
     public String getTelefono() {
         return telefono;
