+++ src/main/java/appFacturacion/Credito.java
@@ -8,11 +8,13 @@ public class Credito extends Pago{
     private String noTarjetaCredito;
     private TipoTarjeta tipoTarejta;
 
-    public Credito(String noTarjetaCredito, TipoTarjeta tipoTarejta, double monto, Date fecha, DetalleOrden detalleOrden, Pago pago, Cliente client, String producto, double Precio, double cantidad, ArrayList<Double> muestra) {
+    public Credito(String noTarjetaCredito, TipoTarjeta tipoTarejta, double monto, Date fecha, DetalleOrden detalleOrden, Pago pago, Cliente client, String producto, float Precio, float cantidad, ArrayList<Double> muestra) {
         super(monto, fecha, detalleOrden, pago, client, producto, Precio, cantidad, muestra);
         this.noTarjetaCredito = noTarjetaCredito;
         this.tipoTarejta = tipoTarejta;
     }
+
+    
 
     public String getNoTarjetaCredito() {
         return noTarjetaCredito;
