+++ src/main/java/appFacturacion/Ocasional.java
@@ -16,10 +16,12 @@ public class Ocasional extends Cliente{
     
     private String direccion;
 
-    public Ocasional(String direccion, String Nombre, String Sexo, Date fecha, DetalleOrden detalleOrden, Pago pago, Cliente client, String producto, double Precio, double cantidad, ArrayList<Double> muestra) {
+    public Ocasional(String direccion, String Nombre, String Sexo, Date fecha, DetalleOrden detalleOrden, Pago pago, Cliente client, String producto, float Precio, float cantidad, ArrayList<Double> muestra) {
         super(Nombre, Sexo, fecha, detalleOrden, pago, client, producto, Precio, cantidad, muestra);
         this.direccion = direccion;
     }
+
+    
 
     public String getDireccion() {
         return direccion;
